refactor: export index로 변경
