+++ components/organisms/UserProfile/IsCoupleProfile.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { calculatingDDay } from 'utils';
-import { IUserProfileProps } from 'types/couple';
+import { IUserProfileProps } from 'types';
 import * as S from './UserProfile.styles';
 
 export function IsCoupleProfile({
